build1

+++ Assets/scripts/HealthMana/ManaBar.cs
@@ -9,13 +9,13 @@ public class ManaBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void SetMana(int mana)
+    public void SetMana(float mana)
     {
         slider.value = mana;
 
         fill.color = gradient.Evaluate(slider.normalizedValue);
     }
-    public void SetMaxMana(int mana)
+    public void SetMaxMana(float mana)
     {
         slider.maxValue = mana;
         slider.value = mana;
